Move the Artifact trait to wasmer_vm
This will allow us to maintain a reference to it inside the instance.
