+++ lib/api/src/sys/tunables.rs
@@ -4,11 +4,10 @@ use std::ptr::NonNull;
 use std::sync::Arc;
 use target_lexicon::PointerWidth;
 use wasmer_compiler::Target;
-use wasmer_engine::Tunables;
 use wasmer_vm::MemoryError;
 use wasmer_vm::{
-    LinearMemory, LinearTable, Memory, MemoryStyle, Table, TableStyle, VMMemoryDefinition,
-    VMTableDefinition,
+    LinearMemory, LinearTable, Memory, MemoryStyle, Table, TableStyle, Tunables,
+    VMMemoryDefinition, VMTableDefinition,
 };
 
 /// Tunable parameters for WebAssembly compilation.
