+++ lib/engine/src/resolver.rs
@@ -1,90 +1,16 @@
 //! Define the `Resolver` trait, allowing custom resolution for external
 //! references.
 
-use crate::{Engine, Export, ExportFunctionMetadata, ImportError, LinkError};
+use crate::{Engine, ImportError, LinkError};
 use more_asserts::assert_ge;
 use wasmer_types::entity::{BoxedSlice, EntityRef, PrimaryMap};
 use wasmer_types::{EntityCounts, ExternType, FunctionIndex, MemoryType, TableType};
 
 use wasmer_vm::{
-    FunctionBodyPtr, ImportFunctionEnv, Imports, MemoryStyle, VMFunctionBody,
-    VMFunctionEnvironment, VMFunctionImport, VMFunctionKind, VMGlobalImport, VMImport,
-    VMImportType, VMMemoryImport, VMTableImport,
+    Export, ExportFunctionMetadata, FunctionBodyPtr, ImportFunctionEnv, Imports, MemoryStyle,
+    NamedResolver, Resolver, VMFunctionBody, VMFunctionEnvironment, VMFunctionImport,
+    VMFunctionKind, VMGlobalImport, VMImport, VMImportType, VMMemoryImport, VMTableImport,
 };
-
-/// Import resolver connects imports with available exported values.
-pub trait Resolver {
-    /// Resolves an import a WebAssembly module to an export it's hooked up to.
-    ///
-    /// The `index` provided is the index of the import in the wasm module
-    /// that's being resolved. For example 1 means that it's the second import
-    /// listed in the wasm module.
-    ///
-    /// The `module` and `field` arguments provided are the module/field names
-    /// listed on the import itself.
-    ///
-    /// # Notes:
-    ///
-    /// The index is useful because some WebAssembly modules may rely on that
-    /// for resolving ambiguity in their imports. Such as:
-    /// ```ignore
-    /// (module
-    ///   (import "" "" (func))
-    ///   (import "" "" (func (param i32) (result i32)))
-    /// )
-    /// ```
-    fn resolve(&self, _index: u32, module: &str, field: &str) -> Option<Export>;
-}
-
-/// Import resolver connects imports with available exported values.
-///
-/// This is a specific subtrait for [`Resolver`] for those users who don't
-/// care about the `index`, but only about the `module` and `field` for
-/// the resolution.
-pub trait NamedResolver {
-    /// Resolves an import a WebAssembly module to an export it's hooked up to.
-    ///
-    /// It receives the `module` and `field` names and return the [`Export`] in
-    /// case it's found.
-    fn resolve_by_name(&self, module: &str, field: &str) -> Option<Export>;
-}
-
-// All NamedResolvers should extend `Resolver`.
-impl<T: NamedResolver> Resolver for T {
-    /// By default this method will be calling [`NamedResolver::resolve_by_name`],
-    /// dismissing the provided `index`.
-    fn resolve(&self, _index: u32, module: &str, field: &str) -> Option<Export> {
-        self.resolve_by_name(module, field)
-    }
-}
-
-impl<T: NamedResolver> NamedResolver for &T {
-    fn resolve_by_name(&self, module: &str, field: &str) -> Option<Export> {
-        (**self).resolve_by_name(module, field)
-    }
-}
-
-impl NamedResolver for Box<dyn NamedResolver + Send + Sync> {
-    fn resolve_by_name(&self, module: &str, field: &str) -> Option<Export> {
-        (**self).resolve_by_name(module, field)
-    }
-}
-
-impl NamedResolver for () {
-    /// Always returns `None`.
-    fn resolve_by_name(&self, _module: &str, _field: &str) -> Option<Export> {
-        None
-    }
-}
-
-/// `Resolver` implementation that always resolves to `None`. Equivalent to `()`.
-pub struct NullResolver {}
-
-impl Resolver for NullResolver {
-    fn resolve(&self, _idx: u32, _module: &str, _field: &str) -> Option<Export> {
-        None
-    }
-}
 
 fn is_compatible_table(ex: &TableType, im: &TableType) -> bool {
     (ex.ty == wasmer_types::Type::FuncRef || ex.ty == im.ty)
@@ -277,94 +203,4 @@ pub fn resolve_imports(
         memory_imports,
         global_imports,
     ))
-}
-
-/// A [`Resolver`] that links two resolvers together in a chain.
-pub struct NamedResolverChain<A: NamedResolver + Send + Sync, B: NamedResolver + Send + Sync> {
-    a: A,
-    b: B,
-}
-
-/// A trait for chaining resolvers together.
-///
-/// ```
-/// # use wasmer_engine::{ChainableNamedResolver, NamedResolver};
-/// # fn chainable_test<A, B>(imports1: A, imports2: B)
-/// # where A: NamedResolver + Sized + Send + Sync,
-/// #       B: NamedResolver + Sized + Send + Sync,
-/// # {
-/// // override duplicates with imports from `imports2`
-/// imports1.chain_front(imports2);
-/// # }
-/// ```
-pub trait ChainableNamedResolver: NamedResolver + Sized + Send + Sync {
-    /// Chain a resolver in front of the current resolver.
-    ///
-    /// This will cause the second resolver to override the first.
-    ///
-    /// ```
-    /// # use wasmer_engine::{ChainableNamedResolver, NamedResolver};
-    /// # fn chainable_test<A, B>(imports1: A, imports2: B)
-    /// # where A: NamedResolver + Sized + Send + Sync,
-    /// #       B: NamedResolver + Sized + Send + Sync,
-    /// # {
-    /// // override duplicates with imports from `imports2`
-    /// imports1.chain_front(imports2);
-    /// # }
-    /// ```
-    fn chain_front<U>(self, other: U) -> NamedResolverChain<U, Self>
-    where
-        U: NamedResolver + Send + Sync,
-    {
-        NamedResolverChain { a: other, b: self }
-    }
-
-    /// Chain a resolver behind the current resolver.
-    ///
-    /// This will cause the first resolver to override the second.
-    ///
-    /// ```
-    /// # use wasmer_engine::{ChainableNamedResolver, NamedResolver};
-    /// # fn chainable_test<A, B>(imports1: A, imports2: B)
-    /// # where A: NamedResolver + Sized + Send + Sync,
-    /// #       B: NamedResolver + Sized + Send + Sync,
-    /// # {
-    /// // override duplicates with imports from `imports1`
-    /// imports1.chain_back(imports2);
-    /// # }
-    /// ```
-    fn chain_back<U>(self, other: U) -> NamedResolverChain<Self, U>
-    where
-        U: NamedResolver + Send + Sync,
-    {
-        NamedResolverChain { a: self, b: other }
-    }
-}
-
-// We give these chain methods to all types implementing NamedResolver
-impl<T: NamedResolver + Send + Sync> ChainableNamedResolver for T {}
-
-impl<A, B> NamedResolver for NamedResolverChain<A, B>
-where
-    A: NamedResolver + Send + Sync,
-    B: NamedResolver + Send + Sync,
-{
-    fn resolve_by_name(&self, module: &str, field: &str) -> Option<Export> {
-        self.a
-            .resolve_by_name(module, field)
-            .or_else(|| self.b.resolve_by_name(module, field))
-    }
-}
-
-impl<A, B> Clone for NamedResolverChain<A, B>
-where
-    A: NamedResolver + Clone + Send + Sync,
-    B: NamedResolver + Clone + Send + Sync,
-{
-    fn clone(&self) -> Self {
-        Self {
-            a: self.a.clone(),
-            b: self.b.clone(),
-        }
-    }
 }
